move the cainjector controller to its own directory
